Added house build, prints, and addGroupNumber to player
I found that we never add a group number if the player buys all the properties in a group. So i added logic to the buyOwnableSquare method in the Game class to handle that. I also made it so the 'house' command works. and added all the view print statements for it.