Added Jail Logic
I added conditions whenever a player rolls that checks if the player is in jail and perfoms actions based on that. The player is only in jail for 3 turns then gets charged $50 and moves on their 3rd turn. Unless doubles are rolled then they get to move. The player in jail still gets to roll to see if they get doubles. This action actually triggers all of the jail logic. I also added a check whenever a player moves, if they land on Go To Jail, it sets the players InJail and moves them to jail